% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping_functions.R
\name{read_planetary_data}
\alias{read_planetary_data}
\title{Read in a single planetary system from yaml}
\usage{
read_planetary_data(yaml_path)
}
\arguments{
\item{yaml_path}{character string providing the path to the yaml file or
a url to the file.}
}
\value{
The final returned object is a list of various planetary properties
with the end value in each case being a tibble. All sourceable values
produce both a value and a \verb{source_\{value\}} variable. Missing values for
source variables indicate noncanon data.
}
\description{
This function reads in a single planetary system file using the syntax of
MekHQ.
}
\examples{

# read in Earth's data directly from GitHub
terra <- read_planetary_data("https://raw.githubusercontent.com/MegaMek/mekhq/refs/heads/master/MekHQ/data/universe/planetary_systems/canon_systems/Terra.yml")

}
