% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping_functions.R
\name{write_planetary_data}
\alias{write_planetary_data}
\title{Write a single planetary system in memory to a file in yaml format}
\usage{
write_planetary_data(planetary_system, path)
}
\arguments{
\item{planetary_system}{A planetary system object created by \code{\link[=read_planetary_data]{read_planetary_data()}}.}

\item{path}{character string providing the path to the yaml file output.}
}
\description{
This function writes a single planetary system object read in by
\code{\link[=read_planetary_data]{read_planetary_data()}}.
}
\examples{

# read in Earth's data directly from GitHub
terra <- read_planetary_data("https://raw.githubusercontent.com/MegaMek/mekhq/refs/heads/master/MekHQ/data/universe/planetary_systems/canon_systems/Terra.yml")
write_planetary_data(terra, "test.yml")

}
